Rename Skeleton, add some code coments
