+++ src/main/java/dev/slimevr/poserecorder/PoseFrameSkeleton.java
@@ -2,7 +2,7 @@ package dev.slimevr.poserecorder;
 
 import dev.slimevr.VRServer;
 import dev.slimevr.vr.processor.ComputedHumanPoseTracker;
-import dev.slimevr.vr.processor.skeleton.SimpleSkeleton;
+import dev.slimevr.vr.processor.skeleton.HumanSkeleton;
 import dev.slimevr.vr.processor.skeleton.SkeletonConfigValue;
 import dev.slimevr.vr.trackers.Tracker;
 
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Map;
 
 
-public class PoseFrameSkeleton extends SimpleSkeleton {
+public class PoseFrameSkeleton extends HumanSkeleton {
 
 	private int frameCursor = 0;
