+++ src/main/java/dev/slimevr/vr/processor/HumanPoseProcessor.java
@@ -2,8 +2,8 @@ package dev.slimevr.vr.processor;
 
 import dev.slimevr.VRServer;
 import dev.slimevr.util.ann.VRServerThread;
+import dev.slimevr.vr.processor.skeleton.Skeleton;
 import dev.slimevr.vr.processor.skeleton.HumanSkeleton;
-import dev.slimevr.vr.processor.skeleton.SimpleSkeleton;
 import dev.slimevr.vr.processor.skeleton.SkeletonConfig;
 import dev.slimevr.vr.processor.skeleton.SkeletonConfigValue;
 import dev.slimevr.vr.trackers.*;
@@ -18,8 +18,8 @@ public class HumanPoseProcessor {
 
 	private final VRServer server;
 	private final List<ComputedHumanPoseTracker> computedTrackers = new FastList<>();
-	private final List<Consumer<HumanSkeleton>> onSkeletonUpdated = new FastList<>();
-	private HumanSkeleton skeleton;
+	private final List<Consumer<Skeleton>> onSkeletonUpdated = new FastList<>();
+	private Skeleton skeleton;
 
 	public HumanPoseProcessor(VRServer server, HMDTracker hmd) {
 		this.server = server;
@@ -105,12 +105,12 @@ public class HumanPoseProcessor {
 			);
 	}
 
-	public HumanSkeleton getSkeleton() {
+	public Skeleton getSkeleton() {
 		return skeleton;
 	}
 
 	@VRServerThread
-	public void addSkeletonUpdatedCallback(Consumer<HumanSkeleton> consumer) {
+	public void addSkeletonUpdatedCallback(Consumer<Skeleton> consumer) {
 		onSkeletonUpdated.add(consumer);
 		if (skeleton != null)
 			consumer.accept(skeleton);
@@ -165,8 +165,8 @@ public class HumanPoseProcessor {
 	@VRServerThread
 	private void updateSekeltonModel() {
 		disconnectAllTrackers();
-		skeleton = new SimpleSkeleton(server, computedTrackers);
-		for (Consumer<HumanSkeleton> sc : onSkeletonUpdated)
+		skeleton = new HumanSkeleton(server, computedTrackers);
+		for (Consumer<Skeleton> sc : onSkeletonUpdated)
 			sc.accept(skeleton);
 	}
