+++ src/main/java/dev/slimevr/vr/trackers/TrackerRole.java
@@ -1,5 +1,12 @@
 package dev.slimevr.vr.trackers;
 
+/**
+ * The tracker role classifies the position and the role of a tracker on user's
+ * body or playspace (like CAMERA or BEACON). Tracker roles are hints for
+ * interacting programs what the tracker means, and they do not correspond to
+ * body's bones on purpose. Example: virtual vive trackers for SteamVR vs actual
+ * SlimeVR trackers.
+ */
 public enum TrackerRole {
 
 	NONE(0, "", "", null),
